+++ database/coingecko_test.go
@@ -2,7 +2,7 @@ package database_test
 
 import (
 	dbtypes "github.com/forbole/bdjuno/database/types"
-	api "github.com/forbole/bdjuno/x/coingecko/apiTypes"
+	api "github.com/forbole/bdjuno/x/supply/coinGeckoTypes"
 )
 
 func (suite *DbTestSuite) TestBigDipperDb_SaveTokenPrice() {
@@ -35,7 +35,7 @@ func (suite *DbTestSuite) TestBigDipperDb_SaveTokenPrice() {
 	var rows []dbtypes.TokenPriceRow
 	err = suite.database.Sqlx.Select(&rows, `SELECT denom,current_price,market_cap,height FROM token_values`)
 	suite.Require().NoError(err)
-	for i,row:=range(rows){
-	suite.Require().True(expected[i].Equals(row))
+	for i, row := range rows {
+		suite.Require().True(expected[i].Equals(row))
 	}
 }
